@@ -1,35 +1,35 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { NgRedux } from 'ng2-redux';
 import { IAppState } from '../../store';
 import { AuthService } from './../auth.service';
 import { UsersActions } from '../../store/users/users.actions';
+import { Store } from '@ngrx/store';
 
-@Component ({
+@Component({
   selector: 'navbar',
-  templateUrl: './navbar.component.html'
+  templateUrl: './navbar.component.html',
 })
 export class NavbarComponent implements OnInit {
   authenticated = false;
-  username: string = null;
+  username: string = '';
 
-  constructor (
-    private ngRedux: NgRedux<IAppState>,
+  constructor(
+    private ngRedux: Store<IAppState>,
     private router: Router,
     private usersActions: UsersActions,
     private authService: AuthService
   ) {}
 
-  ngOnInit () {
+  ngOnInit() {
     this.ngRedux
-      .select(state => state.users)
-      .subscribe(users => {
+      .select((state) => state.users)
+      .subscribe((users: any) => {
         this.authenticated = users.userAuthenticated;
         this.username = users.username;
       });
   }
 
-  logout () {
+  logout() {
     this.usersActions.logout();
     this.authService.deauthenticateUser();
     this.authService.removeUser();
