@@ -3,5 +3,5 @@ export interface ICoreState {
 }
 
 export const initialState: ICoreState = {
-  message: null
+  message: '',
 };
