@@ -1,7 +1,7 @@
-import { MatchesService } from '../../matches/matches.service';
 import { Injectable } from '@angular/core';
-import { NgRedux } from 'ng2-redux';
 import { IAppState } from '..';
+import { Store } from '@ngrx/store';
+import { MatchesService } from '../../matches/matches.service';
 
 export const MATCHES_LOAD = 'matches/LOAD';
 export const MESSAGES_LOAD = 'messages/LOAD';
@@ -10,38 +10,39 @@ export const MESSAGES_LOAD = 'messages/LOAD';
 export class MatchesActions {
   constructor(
     private matchesService: MatchesService,
-    private ngRedux: NgRedux<IAppState>
+    private ngRedux: Store<IAppState>
   ) {}
 
   fetchMatches() {
-    this.matchesService.getMyMatches()
-      .subscribe(matches => {
-        this.ngRedux.dispatch({
-          type: MATCHES_LOAD,
-          matches
-        });
+    this.matchesService.getMyMatches().subscribe((matches: any) => {
+      this.ngRedux.dispatch({
+        type: MATCHES_LOAD,
+        matches,
       });
-  }
-
-  fetchMessages(id) {
-    this.matchesService.getMessages(id).subscribe(response => {
-      if (response.success !== false) {
-        this.ngRedux.dispatch({
-          type: MESSAGES_LOAD,
-          messages: response
-        });
-      } else {
-        console.log(response); // Render error message
-      }
     });
   }
 
-  sendMessage (id, message) {
-    this.matchesService.sendMessage(id, message).subscribe(response => {
+  fetchMessages(id: any) {
+    this.matchesService
+      .getMessages(id)
+      .subscribe((response: { success: boolean }) => {
+        if (response.success !== false) {
+          this.ngRedux.dispatch({
+            type: MESSAGES_LOAD,
+            messages: response,
+          });
+        } else {
+          console.log(response); // Render error message
+        }
+      });
+  }
+
+  sendMessage(id: any, message: any) {
+    this.matchesService.sendMessage(id, message).subscribe((response: any) => {
       if (response.success !== false) {
         this.ngRedux.dispatch({
           type: MESSAGES_LOAD,
-          messages: response
+          messages: response,
         });
       } else {
         console.log(response); // Render error message
