@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { UsersService } from '../../users/users.service';
-import { NgRedux } from 'ng2-redux';
 import { IAppState } from '..';
+import { Store } from '@ngrx/store';
 
 export const USER_LOGGED_IN = 'users/LOGIN';
 export const USER_REGISTERED = 'users/REGISTER';
@@ -12,33 +12,30 @@ export const LOAD_PROFILE = 'profile/LOAD';
 
 @Injectable()
 export class UsersActions {
-  constructor (
-      private usersService: UsersService,
-      private ngRedux: NgRedux<IAppState>
-    ) {}
+  constructor(
+    private usersService: UsersService,
+    private ngRedux: Store<IAppState>
+  ) {}
 
-  register (user) {
-    this.usersService
-      .register(user)
-      .subscribe(result => {
-        this.ngRedux.dispatch({
-            type: USER_REGISTERED,
-            result
-        });
+  register(user: any) {
+    this.usersService.register(user).subscribe((result) => {
+      this.ngRedux.dispatch({
+        type: USER_REGISTERED,
+        result,
       });
+    });
   }
 
-  updateProfile (user) {
-    return this.usersService
-      .editProfile(user);
+  updateProfile(user: any) {
+    return this.usersService.editProfile(user);
   }
 
-  getProfile (id?) {
-    this.usersService.getProfile(id).subscribe(response => {
+  getProfile(id?: any) {
+    this.usersService.getProfile(id).subscribe((response) => {
       if (response.success !== false) {
         this.ngRedux.dispatch({
           type: LOAD_PROFILE,
-          profile: response
+          profile: response,
         });
       } else {
         console.error(response.message);
@@ -46,34 +43,31 @@ export class UsersActions {
     });
   }
 
-  login (user) {
-      this.usersService
-        .login(user)
-        .subscribe(result => {
-          this.ngRedux.dispatch({
-             type: USER_LOGGED_IN,
-             result
-          });
-        });
-  }
-
-  logout () {
-    this.ngRedux.dispatch({
-      type: USER_LOGOUT
+  login(user: any) {
+    this.usersService.login(user).subscribe((result) => {
+      this.ngRedux.dispatch({
+        type: USER_LOGGED_IN,
+        result,
+      });
     });
   }
 
-  getUsersList (query?) {
-    this.usersService.list(query)
-      .subscribe(users => {
-        if (users.success !== false) {
-          this.ngRedux.dispatch({
-            type: USERS_LIST,
-            users
-          });
-        } else {
-          console.log(users); // TODO: Render on DOM
-        }
-      });
+  logout() {
+    this.ngRedux.dispatch({
+      type: USER_LOGOUT,
+    });
+  }
+
+  getUsersList(query?: any) {
+    this.usersService.list(query).subscribe((users) => {
+      if (users.success !== false) {
+        this.ngRedux.dispatch({
+          type: USERS_LIST,
+          users,
+        });
+      } else {
+        console.log(users); // TODO: Render on DOM
+      }
+    });
   }
 }
