@@ -1,37 +1,44 @@
 import { initialState } from './users.state';
 
-import { LOAD_PROFILE, USER_LOGGED_IN, USER_LOGOUT, USER_PROFILE, USER_REGISTERED, USERS_LIST } from './users.actions';
+import {
+  LOAD_PROFILE,
+  USER_LOGGED_IN,
+  USER_LOGOUT,
+  USER_PROFILE,
+  USER_REGISTERED,
+  USERS_LIST,
+} from './users.actions';
 
-function userLogin(state, action) {
+function userLogin(state: any, action: any) {
   const result = action.result;
 
   if (result.success) {
     return Object.assign({}, state, {
       userAuthenticated: result.success,
       token: result.token,
-      username: result.user.email
+      username: result.user.email,
     });
   }
 
   return state;
 }
 
-function userRegistration(state, action) {
+function userRegistration(state: any, action: any) {
   const result = action.result;
   return Object.assign({}, state, {
-    userRegistered: result.success
+    userRegistered: result.success,
   });
 }
 
-function logout(state, action) {
+function logout(state: any, action: any) {
   return Object.assign({}, state, {
     userAuthenticated: false,
     token: null,
-    username: null
+    username: null,
   });
 }
 
-// function profile(state, action) {
+// function profile(state: any, action: any) {
 //   console.log(action);
 //   const result = action.result;
 //   console.log(result.updated.name);
@@ -41,29 +48,29 @@ function logout(state, action) {
 //   });
 // }
 
-function loadProfile(state, profile) {
+function loadProfile(state: any, profile: any) {
   return Object.assign({}, state, { profile });
 }
 
-function loadUsers(state, usersList) {
+function loadUsers(state: any, usersList: any) {
   return Object.assign({}, state, { usersList });
 }
 
-export function usersReducer(state = initialState, action) {
-    switch (action.type) {
-      case USER_REGISTERED:
-        return userRegistration(state, action);
-      case USER_LOGGED_IN:
-        return userLogin(state, action);
-      case USER_LOGOUT:
-        return logout(state, action);
-      // case USER_PROFILE:
-      //   return profile(state, action);
-      case USERS_LIST:
-        return loadUsers(state, action.users);
-      case LOAD_PROFILE:
-        return loadProfile(state, action.profile);
-      default:
-        return state;
-    }
+export function usersReducer(state = initialState, action: any) {
+  switch (action.type) {
+    case USER_REGISTERED:
+      return userRegistration(state, action);
+    case USER_LOGGED_IN:
+      return userLogin(state, action);
+    case USER_LOGOUT:
+      return logout(state, action);
+    // case USER_PROFILE:
+    //   return profile(state, action);
+    case USERS_LIST:
+      return loadUsers(state, action.users);
+    case LOAD_PROFILE:
+      return loadProfile(state, action.profile);
+    default:
+      return state;
+  }
 }
