@@ -1,9 +1,10 @@
 export class ProfileModel {
-    constructor (
-  public name?: string,
-  public age?: number,
-  public location?: string,
-  public gender?: string,
-  public interested?: string
-) {}
+  constructor(
+    public name?: string,
+    public age?: number,
+    public location?: string,
+    public gender?: string,
+    public email?: string,
+    public interested?: string
+  ) {}
 }
