@@ -1,36 +1,38 @@
 import { ProfileModel } from './profile-user.model';
 import { Component, OnInit } from '@angular/core';
 
-import { NgRedux } from 'ng2-redux';
 import { IAppState } from '../../store';
 import { UsersActions } from '../../store/users/users.actions';
 
 import { ActivatedRoute } from '@angular/router';
+import { Store } from '@ngrx/store';
 
 @Component({
   selector: 'profile',
-  templateUrl: './profile.component.html'
+  templateUrl: './profile.component.html',
 })
 export class ProfileComponent implements OnInit {
-  private id: string;
-  private profile: ProfileModel;
+  public id!: string;
+  public profile!: ProfileModel;
 
   constructor(
-    private ngRedux: NgRedux<IAppState>,
+    private ngRedux: Store<IAppState>,
     private usersActions: UsersActions,
     private route: ActivatedRoute
   ) {}
 
   ngOnInit() {
-    this.route.params.subscribe(params => {
+    this.route.params.subscribe((params) => {
       this.id = params['id'];
 
       this.fetch();
     });
 
-    this.ngRedux.select(state => state.users).subscribe(usersState => {
-      this.profile = usersState.profile;
-    });
+    this.ngRedux
+      .select((state) => state.users)
+      .subscribe((usersState: any) => {
+        this.profile = usersState.profile;
+      });
   }
 
   fetch() {
